refactor(client): extract ConnectionStatusBadge from P2PTransfer
